add implementation for adding multi claims

+++ cai/starling.py
@@ -21,9 +21,12 @@ import os
 from cai.core import CaiClaimBlock
 from cai.core import CaiClaimCMSSignature
 from cai.core import CaiStore
+from cai.core import CaiClaim
+from cai.core import CaiAssertionStore
 from cai.jumbf import App11Box
 
 from cai.core import insert_xmp_key
+from cai.core import xmp_read_key
 from cai.jumbf import create_codestream_superbox
 from cai.jumbf import create_json_superbox
 from cai.jumbf import json_to_bytes
@@ -134,11 +137,66 @@ def main():
         with open(args.inject, 'rb') as f:
             raw_bytes = f.read()
 
-        fpath = fname + '-cai' + fext
-        with open(fpath, 'wb') as f:
-            data_bytes = raw_bytes[0:2] + cai_segment.convert_bytes() + raw_bytes[2:]
-            cai_data_bytes = insert_xmp_key(data_bytes, store_label=store_label)
-            f.write(cai_data_bytes)
+            count = 1
+            cai_exist = False
+            for i in range(len(raw_bytes)):
+
+                if raw_bytes[i] == 0xff and raw_bytes[i+1] == 0xeb:
+                    size_hex = (hex(raw_bytes[count+1]).split('x')[-1] + hex(raw_bytes[count+2]).split('x')[-1])
+                    print(size_hex)
+                    size = int('0x' + size_hex, 0)
+                    print(size)
+                    point = count
+                    cai_exist = True
+                    pass
+
+                else:
+                    count+=1
+
+            if cai_exist:
+                print('yes 1')
+                end_position = point+size
+                print('end_position', end_position)
+                parent_claim = xmp_read_key(raw_bytes)
+                print(parent_claim)
+                
+                # assertions_store = CaiAssertionStore(assertions)
+                # claim = CaiClaim(assertions_store, recorder=recorder, parent_claim=parent_claim)
+                # print('claim', claim.convert_bytes())
+                 
+                # cai_store = CaiStore(label=store_label, assertions=assertions, recorder=recorder, key=key, sig=sig, parent_claim=parent_claim)
+                # print(cai_store.convert_bytes())
+                cai_claim_block = CaiClaimBlock()
+                cai_claim_block.content_boxes.append(cai_store)
+                cai_segment = App11Box()
+                cai_segment.payload = cai_claim_block.convert_bytes()
+                size_hex = cai_segment.convert_bytes()[2:4]
+                print(size_hex)
+                size_new = int.from_bytes(size_hex, 'big')
+                print(size_new)
+
+                total = size + size_new
+                print(total)
+                new_size_hex = (total).to_bytes(2, 'big')
+                print(new_size_hex)
+
+                with open(fname + fext, 'rb+') as f:
+                    f.seek(point+1)
+                    f.write(new_size_hex)
+
+                fpath = fname + '-cai' + fext
+                with open(fpath, 'wb') as f:
+                    print('cai segment', cai_segment.convert_bytes())
+                    data_bytes = raw_bytes[0:end_position+1] + cai_segment.convert_bytes() + raw_bytes[end_position+1:]
+                    cai_data_bytes = insert_xmp_key(data_bytes, store_label=store_label)                        
+                    f.write(cai_data_bytes)
+
+            else:           
+                    fpath = fname + '-cai' + fext
+                    with open(fpath, 'wb') as f:
+                        data_bytes = raw_bytes[0:2] + cai_segment.convert_bytes() + raw_bytes[2:]
+                        cai_data_bytes = insert_xmp_key(data_bytes, store_label=store_label)
+                        f.write(cai_data_bytes)
 
 
 if __name__ == "__main__":
